+++ src/Screen/WebErrorScreenEvents.php
@@ -8,30 +8,30 @@ namespace Kuria\Error\Screen;
 abstract class WebErrorScreenEvents
 {
     /**
-     * Emitted when rendering an exception.
+     * Emitted when rendering an exception
      *
      * @param array $view
      */
     const RENDER = 'render';
 
     /**
-     * Emitted when rendering an exception in debug mode.
+     * Emitted when rendering an exception in debug mode
      *
      * @param array $view
      */
     const RENDER_DEBUG = 'render.debug';
 
     /**
-     * Emitted when layout's CSS content is being rendered.
+     * Emitted when layout's CSS content is being output
      *
-     * @param array $view
+     * @param bool $debug
      */
-    const LAYOUT_CSS = 'layout.css';
+    const CSS = 'css';
 
     /**
-     * Emitted when layout's JS content is being rendered.
+     * Emitted when layout's JS content is being output
      *
-     * @param array $view
+     * @param bool $debug
      */
-    const LAYOUT_JS = 'layout.js';
+    const JS = 'js';
 }